requirements.txt
Flask==3.0.0
flask-restful==0.3.10
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
requests==2.31.0
Faker==20.1.0
ollama
